v2/guest: Move `Argv` to `call::types`
The type will be shared with `gdbcall`, therefore provide an
implementation for argv of length of 4.
