+++ src/v2/src/guest/call/syscall/fcntl.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
+use super::super::types::Argv;
 use super::super::Alloc;
-use super::types::Argv;
 use super::PassthroughAlloc;
 use crate::guest::alloc::Allocator;
 use crate::guest::call::alloc::kind;
