+++ src/v2/src/guest/call/syscall/types/mod.rs
@@ -3,14 +3,12 @@
 //! Syscall-specific types.
 
 mod alloc;
-mod argv;
 mod bytes;
 mod result;
 mod sockaddr;
 mod sockopt;
 
 pub use alloc::*;
-pub use argv::*;
 pub use bytes::*;
 pub use result::*;
 pub use sockaddr::*;
