+++ src/v2/src/guest/call/syscall/write.rs
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
-use super::types::{Argv, StagedBytesInput};
+use super::super::types::Argv;
+use super::types::StagedBytesInput;
 use super::Alloc;
 use crate::guest::alloc::{Allocator, Collector, Input};
 use crate::Result;
