+++ src/v2/src/guest/call/types/argv.rs
@@ -5,10 +5,31 @@ use crate::NULL;
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct Argv<const N: usize>(pub [usize; N]);
 
+impl<const N: usize> From<Argv<N>> for [usize; N] {
+    #[inline]
+    fn from(argv: Argv<N>) -> Self {
+        argv.0
+    }
+}
+
+impl From<Argv<0>> for [usize; 4] {
+    #[inline]
+    fn from(_: Argv<0>) -> Self {
+        [NULL, NULL, NULL, NULL]
+    }
+}
+
 impl From<Argv<0>> for [usize; 6] {
     #[inline]
     fn from(_: Argv<0>) -> Self {
         [NULL, NULL, NULL, NULL, NULL, NULL]
+    }
+}
+
+impl From<Argv<1>> for [usize; 4] {
+    #[inline]
+    fn from(argv: Argv<1>) -> Self {
+        [argv.0[0], NULL, NULL, NULL]
     }
 }
 
@@ -19,10 +40,24 @@ impl From<Argv<1>> for [usize; 6] {
     }
 }
 
+impl From<Argv<2>> for [usize; 4] {
+    #[inline]
+    fn from(argv: Argv<2>) -> Self {
+        [argv.0[0], argv.0[1], NULL, NULL]
+    }
+}
+
 impl From<Argv<2>> for [usize; 6] {
     #[inline]
     fn from(argv: Argv<2>) -> Self {
         [argv.0[0], argv.0[1], NULL, NULL, NULL, NULL]
+    }
+}
+
+impl From<Argv<3>> for [usize; 4] {
+    #[inline]
+    fn from(argv: Argv<3>) -> Self {
+        [argv.0[0], argv.0[1], argv.0[2], NULL]
     }
 }
 
@@ -44,12 +79,5 @@ impl From<Argv<5>> for [usize; 6] {
     #[inline]
     fn from(argv: Argv<5>) -> Self {
         [argv.0[0], argv.0[1], argv.0[2], argv.0[3], argv.0[4], NULL]
-    }
-}
-
-impl From<Argv<6>> for [usize; 6] {
-    #[inline]
-    fn from(argv: Argv<6>) -> Self {
-        argv.0
     }
 }
