+++ src/v2/src/guest/call/types/mod.rs
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: Apache-2.0
+
+//! Call-specific types.
+
+mod argv;
+
+pub use argv::*;
